+++ iwopy/core/problem.py
@@ -47,6 +47,7 @@ class Problem(Base, metaclass=ABCMeta):
 
         self._cons_mi = None
         self._cons_ma = None
+        self._cons_tol = None
 
     def var_names_int(self):
         """
@@ -287,12 +288,16 @@ class Problem(Base, metaclass=ABCMeta):
         self.cons.append(constraint)
 
         cmi, cma = constraint.get_bounds()
+        ctol = np.zeros(constraint.n_components(), dtype=np.float64)
+        ctol[:] = constraint.tol
         if self._cons_mi is None:
             self._cons_mi = cmi
             self._cons_ma = cma
+            self._cons_tol = ctol
         else:
             self._cons_mi = np.append(self._cons_mi, cmi, axis=0)
             self._cons_ma = np.append(self._cons_ma, cma, axis=0)
+            self._cons_tol = np.append(self._cons_tol, ctol, axis=0)
 
     @property
     def min_values_constraints(self):
@@ -319,6 +324,19 @@ class Problem(Base, metaclass=ABCMeta):
 
         """
         return self._cons_ma
+
+    @property
+    def constraints_tol(self):
+        """
+        Gets the tolerance values of constraints
+
+        Returns
+        -------
+        ctol : numpy.ndarray
+            The constraint tolerance values, shape: (n_constraints,)
+
+        """
+        return self._cons_tol
 
     @property
     def n_objectives(self):